unicorn_parsing/requirements.txt
layoutparser
Pillow
# detectron2 설치는 특이하므로 아래 명령어를 사용해주세요.
# pip install "detectron2@git+https://github.com/facebookresearch/detectron2.git@v0.6#egg=detectron2" 